Dependency change, UX change
Yad added as a dependency for non EndeavourOS users who don't have yad preinstalled. Changed the way it asked for personal info removal and link generation. The previous way was pretty unclear as not everyone is going to read all that text at the top, leaving the last checkboxes unnoticed.