Welcome app translated to Serbian language (Latin font).
I've tried to convey the actual meaning as accurate as possible, as a result I have left a few words / phrases in English as they either do not translate well or completely loose their meaning. This is due to the fact that IT terms in Serbian language are mostly serbianized versions of their English counterparts.